feat(dashboard): Add UI host to .env.example
