+++ apps/dashboard/.env.example
@@ -1,4 +1,5 @@
-VITE_POSTHOG_HOST=https://us.i.posthog.com
+VITE_POSTHOG_HOST=https://e.yourdomain.com
 VITE_POSTHOG_KEY=phc_xxxxxxxxxxxxxxxxxxxxxxxxxxxxx
+VITE_POSTHOG_UI_HOST=https://us.posthog.com
 VITE_SERVER_URL=http://localhost:9876
