+++ apps/landing-page/.env.example
@@ -1,2 +1,3 @@
-VITE_POSTHOG_HOST=https://us.i.posthog.com
+VITE_POSTHOG_HOST=https://e.yourdomain.com
 VITE_POSTHOG_KEY=phc_xxxxxxxxxxxxxxxxxxxxxxxxxxxxx
+VITE_POSTHOG_UI_HOST=https://us.posthog.com
