+++ apps/landing-page/src/layouts/Landing.astro
@@ -1,7 +1,7 @@
 ---
 import "@packages/ui/globals.css";
 
-import { VITE_POSTHOG_HOST, VITE_POSTHOG_KEY } from "astro:env/client";
+import { VITE_POSTHOG_HOST, VITE_POSTHOG_KEY, VITE_POSTHOG_UI_HOST } from "astro:env/client";
 import type { SupportedLng } from "@packages/localization";
 import { translate } from "@packages/localization";
 import { getAstroPosthogConfig } from "@packages/posthog/client";
@@ -17,6 +17,7 @@ const brandDescription = translate("common.brand.description", { lng: lang });
 const posthogEnv = {
    VITE_POSTHOG_HOST,
    VITE_POSTHOG_KEY,
+   VITE_POSTHOG_UI_HOST,
 };
 ---
