+++ packages/environment/src/client.ts
@@ -4,6 +4,7 @@ import { parseEnv } from "./helpers";
 const EnvSchema = z.object({
    VITE_POSTHOG_HOST: z.string(),
    VITE_POSTHOG_KEY: z.string(),
+   VITE_POSTHOG_UI_HOST: z.string(),
    VITE_SERVER_URL: z.string(),
 });
 export type ClientEnv = z.infer<typeof EnvSchema>;
